+++ inst/examples/app/analytics/app.R
@@ -17,8 +17,11 @@ library(DT)
 library(shiny.telemetry)
 
 # Default storage backend using LogFile
+log_file_path <- file.path("..", "instrumentation", "user_stats.txt")
+if (!dir.exists(dirname(log_file_path))) log_file_path <- "user_stats.txt"
+
 data_storage <- DataStorageLogFile$new(
-  log_file_path = file.path(getwd(), "user_stats.txt")
+  log_file_path = log_file_path
 )
 
 # This sample application includes a configuration for RSConnect deployments,
@@ -32,4 +35,4 @@ if (Sys.getenv("R_CONFIG_ACTIVE") == "rsconnect") {
 
 analytics_app(data_storage = data_storage)
 
-# shiny::shinyAppDir(file.path("inst", "examples", "app", "analytics"))
+# shiny::shinyApp(system.file("examples", "app", "analytics", package = "shiny.telemetry")) # nolint: commented_code, line_length.