+++ inst/examples/mariadb/mariadb_analytics.R
@@ -21,3 +21,5 @@ data_storage <- DataStorageMariaDB$new(
 )
 
 analytics_app(data_storage = data_storage)
+
+# shiny::shinyAppFile(system.file("examples", "mariadb", "mariadb_analytics.R", package = "shiny.telemetry")) # nolint: commented_code, line_length.